reschedule "run once" jobe so unschedule themselves after a week
A job that is not scheduled and should execute only once because of the
execution of cron.run_job() should not be canceled because when it
starts, it is unscheduled. So now its schedule is altered to unschedule
itself after a week. Thus the schedule remains - so the job will not be
canceled - provided that the job finishes within a week.